Added Product endponits -  getProducts, getProcudtById, updateproduct, deleteProduct, createProduct

+++ Backend/src/controllers/User-Controller.ts
@@ -2,10 +2,10 @@ import { RequestHandler } from "express";
 const User = require('../Schema/User');
 
 const getAllUsers: RequestHandler = async (req, res) => {
-    const count = req.query.count
+    const latest = req.query.latest
     try {
-        const retrivedUsers = count ?
-            await User.find().sort({id : -1}).limit(count)
+        const retrivedUsers = latest ?
+            await User.find().sort({createdAt : -1}).limit(10)
             : await User.find();
         res.json( retrivedUsers ).status(200);
     } catch (e) {
@@ -15,6 +15,9 @@ const getAllUsers: RequestHandler = async (req, res) => {
 
 const getUserById :RequestHandler = async (req, res) => {
     const userId = req.params.id
+    if (!userId) {
+        return res.json("User not found").status(404);
+    }
     try {
         const retrivedUser = await User.findOne({id: userId});
         res.json( retrivedUser ).status(200);
@@ -25,7 +28,9 @@ const getUserById :RequestHandler = async (req, res) => {
 
 const updateUser :RequestHandler = async (req, res) => {
     const userId = req.params.id
-
+    if (!userId) {
+        return res.json("User not found").status(404);
+    }
     try {
         const updatedUser = await User.findByIdAndUpdate(userId, {
             $set : req.body
@@ -38,6 +43,9 @@ const updateUser :RequestHandler = async (req, res) => {
 
 const deleteUser :RequestHandler = async (req, res) => {
     const userId = req.params.id
+    if (!userId) {
+        return res.json("User not found").status(404);
+    }
     try {
         const deletedUser = await User.findByIdAndDelete(userId);
         res.json( deletedUser ).status(200);
